% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_features.R
\name{build_features}
\alias{build_features}
\title{Builds the feature-matrix from a text-vector}
\usage{
build_features(x, term_count_min = 1, mdl = NULL, parallel = TRUE,
  quiet = FALSE)
}
\arguments{
\item{x}{a vector of text}

\item{term_count_min}{a number passed to
\code{\link[text2vec]{prune_vocabulary}}, defaults to 1. In case the function
is used for training, it can and should be set to some higher value, i.e., 3.}

\item{mdl}{is a list of existing models-data (containing the vectorizer, the
tfidf, and the lsa object), defaults to NULL, in which case it is rebuild}

\item{parallel}{T/F if the task should be executed in parallel, defaults to TRUE}

\item{quiet}{T/F if the function remains silent, defaults to FALSE}
}
\value{
a list of two: a dgCMatrix that contains the features (columns) for
each text (row) and as a second element a list of the model that can be passed
as mdl
}
\description{
Builds the feature-matrix from a text-vector
}
\examples{
text <- c(
  "This is a first text that describes something",
  "A second Text That USES A LOT of CAPITALS",
  "Lastly MANY!!!! (like, really a lot!) punctuations!!!"
)

build_features(text)

dtm <- c("capit", "someth", "punctuat", "use", "mani", "second", "last",
   "describ", "like", "first", "realli")
build_features(text, dtm = dtm)

# a second example
train <- c("Banking is finance", "flowers are not houses", "finance is power", "houses are build")
test <- c("finance is greed", "flowers belong in the garbage", "houses are build")

a1 <- build_features(test)
a12 <- build_features(test, mdl = a1$mdl)

a2 <- build_features(train, mdl = a1$mdl)
a2$res
}
