% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_troll.R
\name{predict_troll}
\alias{predict_troll}
\title{Detect if given texts are trolls}
\usage{
predict_troll(x, model_ = NULL, mdl_data_ = NULL)
}
\arguments{
\item{x}{a vector of text}

\item{model_}{a model that is passed to predict, defaults to the \code{model}
supplied with this package}

\item{mdl_data_}{a model as returned by \code{\link{build_features}} (the mdl)
containing the vectorizer, tfidf, and the lsa objects. Defaults to the
\code{mdl_data} from this package.}
}
\value{
a vector with the same lengths as x that holds the predicted probabilities
that the given text is trolling
}
\description{
Detect if given texts are trolls
}
\examples{
text <- c("You suck, die!", "What a nice world we have today", "I like you", "I hate you")
(pred <- predict_troll(text))
}
