% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_api.R
\name{run_api}
\alias{run_api}
\title{Run the Plumber API}
\usage{
run_api(port = 8000)
}
\arguments{
\item{...}{parameters passed to \code{\link[plumber]{plumber}}}
}
\value{
invisible NULL
}
\description{
Run the Plumber API
}
\examples{
\dontrun{
  run_api()
  # try to got to: http://127.0.0.1:8000/trollR
  # or use http://127.0.0.1:8000/trollR?text=This may be a troll comment
}
}
