% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Shiny App Launcher}
\usage{
run_shiny(example = "trollR")
}
\arguments{
\item{example}{name of the app, defaults to trollR}
}
\value{
Nothing
}
\description{
Shiny App Launcher
}
\examples{
\dontrun{
 run_shiny()
}
}
