% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shitwordlist}
\alias{shitwordlist}
\title{Shitword list}
\format{A character vector containing 349 words.
\describe{
  \item{test}{this is only for illustration purpose how to document data}
  \item{london}{hello whatup}
}}
\usage{
shitwordlist
}
\description{
A list of Enlish curse words scraped from https://www.noswearing.com/dictionary
}
\keyword{datasets}
